Updated English, Spanish, & Chinese- completed
line 246 in Chinese language- want to make sure I entered the {0} appropriately at the end

In general- Britt added 2 screens that aren't included in this version. 1: the on boarding age restriction pop-up (slides 11-13 on translation slides) and 2: text to appear at the top of the home screen before July (slides 27-29 on translation slides)

In the "My Recorded Walks" screen- in the English version there is a "back" option on the upper left corner of the screen, this isn't present in the Spanish & Chinese versions.